+++ app/src/main/java/trianglz/models/Subcategory.java
@@ -34,7 +34,7 @@ public class Subcategory {
     @SerializedName("grade_items_total")
     public int gradeItemsTotal;
     @SerializedName("grade_items_grade")
-    public int gradeItemsGrade;
+    public double gradeItemsGrade;
     @SerializedName("grade_items")
     public ArrayList<GradeItems> gradeItems;
     @SerializedName("percentage")
